--- a/typed-racket-lib/typed-racket/static-contracts/combinators/any.rkt
+++ b/typed-racket-lib/typed-racket/static-contracts/combinators/any.rkt
@@ -15,7 +15,8 @@
 (provide/cond-contract
  [any/sc static-contract?])
 
-(provide any/sc:)
+(provide any/sc:
+         (rename-out [any-combinator? any/sc?]))
 
 
 ;;Printing