This PR adds dependent function types to Typed Racket. Specifically, it adds the following:

```
* new type DFun (see DFun def in type-rep.rkt)
  + represents a dependent function (i.e. arguments
    depend on one another)
  + syntax is like ->i, but w/o the 'i', e.g.
    (-> ([v : (Vec A)] 
         [i : (v) Nat #:where (< i (vec-len v))])
         A)
    order of dependencies doesn't matter, as long as there's
    no cycle in the dependencies
  + contracts for DFun generate ->i contracts (obviously)

* tc-expr/check can assign existential objects
  + sometimes we want a local/existential object for a value
    (like when applying a dependent function), tc-expr/check
    can assign one (notably this is used by tc-dep-fun-arg)

* #:with-logical-refinements flag turns on more reasoning
  namely existential objects for functions w/ dependent types
  and linear integer information for integer primitives (see below)


* integer arith prims get propositions w/ #:with-logical-refinements
  I ran into a bit of chicken and egg problem since our theory knows
  about certain integer operations (e.g. *, +, -, <, <=, =, >, <=). I
  decided to go ahead and hard-code their logical info so it is
  explicitly added when we observe their application while type checking.
  This means at the moment you couldn't express the same thing with the
  types we allow users to write... trade offs =\

* unsafe-require/typed/provide added
  see prims-contract.rkt

* misc
  + vector-length has a latent object now
  + tc/dep-lambda checks dependent function bodies (tc-lambda-unit.rkt)
  + logical propositions are extracted when recording initial types
    for identifiers (see tc-let-unit.rkt and perhaps other areas)
  + extract-props returns a simplified type if props were extracted
    in addition to any extracted props to avoid infinite loops
    e.g. old behavior: (extract-props obj {x : Int | x < 42})
                       ==> (< obj 42)
         new behavior: (extract-props obj {x : Int | x < 42})
                       ==> (values Int (< obj 42))

* VecLenPE treatment (uninterpreted-PE?)
  + there is special treatment of the `VecLenPE` path element in several places
     (grep for `uninterpreted-PE?`), this is because `VecLenPE` is the first path element
     that does not refer to some sub-portion of a type. If we learn something about
     the VecLenPE of an object, we need to remember it... and so currently we explicitly
     test for this case and store the info. In the future, if there are more such path elements,
     we would need to add these to the `uninterpreted-PE?` predicate as well (or distinguish
     them from interpreted path elements in some other way)
```

TODO
- [x] dependent function documentation
- [x] better documentation for linear integer stuff
- [x] add docs for unsafe-require/typed[/provide]
- [x] add SPECIFIC documentation of which functions get more logical info when the `#:with-logical-refinements` keyword is provided on the `#lang` line