@pnwamk 2017-07-17
The math tests are failing due to some exponential proposition growth. It appears bubbling up the types of existential variables in the then- and else-propositions of type checking results is not a feasible technique (which isn't that surprising... it's pretty hacky). Will have to look at more efficient ways to propagate that information up (mutation? another field in tc-results? ?)