@pnwamk 2017-09-08
we're passing tests and have some preliminary performance stats.

I think I need to write up a summary still and maybe play around with the perf degradation and see if I can narrow down what's causing it.

```
performance results:
schml-specify-rep: 2.2% faster (6.36s to 6.22s)
schml-interp-casts-help: 3.4% slower (15.86s to 16.4s)
parser: 7.04% slower (1.42s to 1.52s)
old-metrics: 4.02% slower (1.74s to 1.81s)
new-metrics: 4.8% slower (2.29s to 2.4s)
math-flonum: 10.11% slower (2.77s to 3.05s)
fsm: 4.29% slower (5.6s to 5.84s)
forth: 5.56% slower (9.72s to 10.26s)
dungeon: 4.9% slower (10.61s to 11.13s)
acquire: 1.48% slower (28.37s to 28.79s)
```

(perf results calculated by [this](https://github.com/pnwamk/tr-performance) code)