@pnwamk 2017-09-11
w/ commit b8d6e05 we get the following timing results (notably the math related file is now more or less unaffected):

```
6.10 vs PR
Compile time ratio (old time / new time):
schml-specify-rep: 1.02 (i.e. faster, 6.33s (σ 0.01) to 6.18s (σ 0.01))
schml-interp-casts-help: 0.99 (i.e. slower, 15.84s (σ 0.02) to 16.0s (σ 0.05))
parser: 0.93 (i.e. slower, 1.42s (σ 0.0) to 1.52s (σ 0.0))
old-metrics: 0.96 (i.e. slower, 1.74s (σ 0.0) to 1.81s (σ 0.0))
new-metrics: 1.0 (i.e. slower, 2.3s (σ 0.0) to 2.3s (σ 0.0))
math-flonum: 0.99 (i.e. slower, 2.77s (σ 0.01) to 2.81s (σ 0.0))
fsm: 0.95 (i.e. slower, 5.58s (σ 0.0) to 5.85s (σ 0.01))
forth: 0.95 (i.e. slower, 9.73s (σ 0.01) to 10.28s (σ 0.02))
dungeon: 0.96 (i.e. slower, 10.56s (σ 0.01) to 10.96s (σ 0.02))
acquire: 0.98 (i.e. slower, 28.33s (σ 0.07) to 28.83s (σ 0.03))


;; 6.2 vs PR
Compile time ratio (old time / new time):
schml-specify-rep: 2.17 (i.e. faster, 13.43s (σ 0.01) to 6.2s (σ 0.01))
schml-interp-casts-help: 9.54 (i.e. faster, 152.69s (σ 1.34) to 16.0s (σ 0.05))
parser: 0.91 (i.e. slower, 1.38s (σ 0.0) to 1.52s (σ 0.0))
old-metrics: 0.99 (i.e. slower, 1.79s (σ 0.0) to 1.81s (σ 0.01))
new-metrics: 1.41 (i.e. faster, 3.25s (σ 0.01) to 2.3s (σ 0.0))
math-flonum: 1.12 (i.e. faster, 3.14s (σ 0.01) to 2.81s (σ 0.0))
fsm: 0.95 (i.e. slower, 5.58s (σ 0.0) to 5.85s (σ 0.02))
forth: 0.97 (i.e. slower, 9.93s (σ 0.03) to 10.23s (σ 0.02))
dungeon: 0.95 (i.e. slower, 10.39s (σ 0.01) to 10.98s (σ 0.04))
acquire: 1.11 (i.e. faster, 31.98s (σ 0.03) to 28.85s (σ 0.04))
```