@pnwamk 2017-09-13
I'm going to add 'fixed' after in-line comments when I address them so I can keep track of what I've done... if there's a better way let me know! =)