@pnwamk 2017-09-13
I've addressed just about every PR comment. There's just a slight bug I need to fix when I get home (it's the failing integration test).