@pnwamk 2017-09-14
@samth I believe the only thing left before this is ready to merge is whether we allow `->` with dependent functions (as far as the user level syntax), or should we require some other arrow (e.g. `->i`, etc), or should we support `->` _and_ other arrows, etc.