@samth 2017-09-14
@rfindler I'm interested in your thoughts on the naming question for writing dependent function types in Typed Racket. There are a few options:

1. Re-use `->` for dependent functions. Since types aren't just expressions, it's possible to have this work fine. This is what @pnwamk has already implemented. Cons: not like contracts.
2. Use `->i` for dependent functions. Pros: aligns with contracts. Cons: the `i` refers to an aspect of the semantics that isn't relevant for type checking, and is thus confusing.
3. Allow `->d` and `->i`, meaning the same thing.
4. Allow `->`, `->d`, and `->i`, all meaning the same thing. 

I'm somewhat leaning toward 4, but I'd like to know what you think as well.