@pnwamk 2017-09-18
Changes from comments have been made, commits are squashed and ready for merge, just waiting for word on additional/different `->` syntax for dependent functions (if any is desired in this PR).