@rfindler 2017-09-18
I think the right approach (which is actually what seems to be being followed, but perhaps not articulated, so maybe not explicitly followed?) is that whenever there is an arbitrary choice in the syntax to be made, follow the contract library (or talk to me to get the contract library extended/changed). So, writing the dependencies explicitly in parens the way it happens above matches, good. The parens are in the same places, good. The `#:where` keyword isn't something that has a direct correspondence in the contract system; well perhaps via `#:pre`: possibly good, possibly bad. I guess the right thing is to think about whether or not I can add something simple to the contract system to match what you have to do, or if you can use `#:pre` instead. Or maybe I add something *and* you support `#:pre`. Etc.

I'm happy to discuss specifics more. 

Meanwhile, I agree that the `i` isn't suitable because there is no such thing as "indy". I'd probably just use `->` only; or I'd separate them and use `i` only as an homage to the contract system. I wouldn't have multiple different ways to spell the same thing and I wouldn't use `d` because that is a different notation in the contract system.