@samth 2017-09-18
Ok, we'll go with plain `->`. @pnwamk what do you think about `#:pre` (it makes sense as a precondition)?