@pnwamk 2017-09-18
A #:pre could definitely express anything that could be written in a
#:where clause.

Having #:pre would probably make the most sense as an _additional_ field on
our representation of dependent function types. This has the downside of
making our dependent function types a little more complicated, but would
probably provide more useful error messages more often for users (i.e. we
could much more easily report when the domain types were satisfied but the
precondition was unprovable at a given application site using a #:pre than
we could if we were merely using #:where clauses as syntactic sugar for a
Refinement type).
On Mon, Sep 18, 2017 at 5:13 PM Sam Tobin-Hochstadt <
notifications@github.com> wrote:

> Ok, we'll go with plain ->. @pnwamk <https://github.com/pnwamk> what do
> you think about #:pre (it makes sense as a precondition)?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/racket/typed-racket/pull/584#issuecomment-330358059>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ADfg6sQ4uMGgK9HSMZuri6CVy32yx_8sks5sjtzwgaJpZM4OZahR>
> .
>