@rfindler 2017-09-18
I don't know what a where is, but my initial reaction is to ask why pre and
where aren't interchangeable (is the answer only that pre allows Racket
expressions and pre is more restricted? If so then I suggest changing the
notation to write "pre" instead of "where" to make things more similar)

Also: is there a grammar somewhere?