@samth 2017-09-18
What would be the distinction between `#:pre` and `#:where` that you're thinking of? And would this distinction make sense for contracts too?